+++ single-maestro.php
@@ -1,0 +1,9 @@
+<?php get_template_part('parts/header'); ?>
+
+<div id="primary" class="content-area">
+    <main id="main" class="site-main" role="main">
+        <?php get_template_part('patterns/template-single-maestro') ?>
+    </main><!-- #main -->
+</div><!-- #primary -->
+
+<?php get_template_part('parts/footer'); ?>