+++ CHANGELOG.md
@@ -5,8 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.3.0] 2025-09-02
 ### Added
+- Add single-post template for `maestro` post-type
 - Add `browser-sync` with command `npm run dev` to support live-reload
 ### Changed
 - `templates/template-single-post-footer.php` moved to `parts/single-post-footer.php`
@@ -178,7 +179,8 @@ the slug's name
 - Initial `index.php`, `style.css`, and `functions.php`
 - This Changelog
 
-[Unreleased]: https://github.com/erremauro/cigno-zen/compare/v1.2.2...HEAD
+[Unreleased]: https://github.com/erremauro/cigno-zen/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/erremauro/cigno-zen/releases/tag/v1.3.0
 [1.2.2]: https://github.com/erremauro/cigno-zen/releases/tag/v1.2.2
 [1.2.1]: https://github.com/erremauro/cigno-zen/releases/tag/v1.2.1
 [1.2.0]: https://github.com/erremauro/cigno-zen/releases/tag/v1.2.0